+++ config/example_options.ini
@@ -29,6 +29,11 @@ Token = bot_token
 Spotify_ClientID =
 Spotify_ClientSecret =
 
+[Cogs]
+# This specify with cogs (group of commands) should be loaded
+# If you have no idea what this does, it's better to leave it as is
+Cogs_Load = help autoplaylist moderate info queue botmanipulate playback dev utility
+
 [Permissions]
 # This option determines which user has full permissions and control of the bot.
 # You can only set one owner, but you can use permissions.ini to give other
